refactor(profiler): move all profiler files to be under devtool-tabs where the initial profiler component is rendered
