+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/profiler.module.ts
