+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/recording/recording-modal/recording-dialog/recording-dialog.component.html
