+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/recording/recording-modal/recording-modal.component.css
@@ -1,4 +1,4 @@
-@import '../../../../../../../../../node_modules/@angular/cdk/overlay-prebuilt.css';
+@import '../../../../../../../../node_modules/@angular/cdk/overlay-prebuilt.css';
 
 :host {
   overflow: hidden;
