+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/recording/recording-modal/recording-modal.component.html
