+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/recording/timeline/record-formatter/webtreegraph-formatter/index.ts
