+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/recording/timeline/timeline.component.html
