+++ projects/ng-devtools/src/lib/devtools-tabs/profiler/recording/timeline/timeline.module.ts
